Moved defvjp wrappers back into core since there are so few of them left and they're handy for defining the vspace grads
