+++ autograd/extend.py
@@ -1,58 +1,5 @@
-from itertools import count
-from .util import subval
-from .core import defvjp_argnums, defjvp, defjvp_argnums, sum_outgrads
-
-# Expose API for extending autograd
-from .tracer import Box, primitive, notrace_primitive, getval
-from .core import SparseObject, VSpace, vspace
-
-# -------------------- reverse mode defgrad wrappers --------------------
-
-def defvjp(fun, *vjpmakers, **kwargs):
-    argnums = kwargs.get('argnums', count())
-    vjps_dict = {argnum : translate_vjp(vjpmaker, fun, argnum)
-                 for argnum, vjpmaker in zip(argnums, vjpmakers)}
-    def vjp_argnums(argnums, ans, args, kwargs):
-        L = len(argnums)
-        # These first two cases are just optimizations
-        if L == 1:
-            argnum = argnums[0]
-            vjp = vjps_dict[argnum](ans, *args, **kwargs)
-            return lambda g: (vjp(g),)
-        elif L == 2:
-            argnum_0, argnum_1 = argnums
-            vjp_0 = vjps_dict[argnum_0](ans, *args, **kwargs)
-            vjp_1 = vjps_dict[argnum_1](ans, *args, **kwargs)
-            return lambda g: (vjp_0(g), vjp_1(g))
-        else:
-            vjps = [vjps_dict[argnum](ans, *args, **kwargs) for argnum in argnums]
-            return lambda g: (vjp(g) for vjp in vjps)
-
-    defvjp_argnums(fun, vjp_argnums)
-
-def translate_vjp(vjpfun, fun, argnum):
-    if vjpfun is None:
-        return lambda ans, *args, **kwargs: lambda g: vspace(args[argnum]).zeros()
-    elif callable(vjpfun):
-        return vjpfun
-    else:
-        raise Exception("Bad VJP '{}' for '{}'".format(vjpfun, fun.__name__))
-
-def defvjp_argnum(fun, vjpmaker):
-    def vjp_argnums(argnums, *args):
-        vjps = [vjpmaker(argnum, *args) for argnum in argnums]
-        return lambda g: (vjp(g) for vjp in vjps)
-    defvjp_argnums(fun, vjp_argnums)
-
-# -------------------- forward mode defgrad wrappers  --------------------
-
-def defjvp_argnum(fun, jvpmaker):
-    def jvp_argnums(argnums, gs, ans, args, kwargs):
-        return sum_outgrads(jvpmaker(argnum, g, ans, args, kwargs)
-                            for argnum, g in zip(argnums, gs))
-    defjvp_argnums(fun, jvp_argnums)
-
-def def_linear(fun):
-    """Flags that a function is linear wrt all args"""
-    defjvp_argnum(fun, lambda argnum, g, ans, args, kwargs:
-                  fun(*subval(args, argnum, g), **kwargs))
+# Exposes API for extending autograd
+from .tracer import Box, primitive, notrace_primitive
+from .core import (SparseObject, VSpace, vspace,
+                   defvjp_argnums, defvjp_argnum, defvjp,
+                   defjvp_argnums, defjvp_argnum, defjvp, def_linear)
